requirements.txt
genanki
osmpythontools
pyosmium
